% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggordi.R
\name{ggordi}
\alias{ggordi}
\title{ggordi}
\usage{
ggordi(
  pcobj,
  choices = 1:2,
  scale = 1,
  sites.geom = "point",
  sites.shape = NULL,
  sites.leg.title.s = NULL,
  sites.fill = NULL,
  sites.leg.title.f = NULL,
  sites.color = NULL,
  sites.leg.title.c = NULL,
  sites.labs = NULL,
  sites.pal = NULL,
  sites.shape.type = "fill",
  sites.size = NULL,
  draw.arrows = T,
  arrow.color = rgb(0, 0, 0, 0.7),
  species.labs = NULL,
  species.geom = "text",
  species.labs.size = 4,
  species.labs.adjust = 0.6,
  ellipse.groups = NULL,
  ellipse = T,
  ellipse.labs = T,
  ellipse.level = 0.69,
  ellipse.show.legend = NA,
  ellipse.type = "norm",
  ellipse.size = 1,
  ellipse.alpha = 0.15,
  lty.pal = NULL,
  axis.lang = "EN",
  draw.axis = T
)
}
\arguments{
\item{pcobj}{an object returned by prcomp (will add others later)}

\item{choices}{which PCs to plot}

\item{scale}{covariance biplot (scale = 1), form biplot (scale = 0). When scale = 1, the inner product between the variables approximates the covariance and the distance between the points approximates the Mahalanobis distance.
scores/sites args}

\item{sites.geom}{either 'point' or 'text'}

\item{sites.shape}{shape aes for sites}

\item{sites.leg.title.s}{legend title for shape}

\item{sites.fill}{fill aes for sites}

\item{sites.leg.title.f}{legend title for fill}

\item{sites.color}{color aes for sites}

\item{sites.leg.title.c}{legend title for color}

\item{sites.labs}{optionnal vector oflabels for sites if different than rownames}

\item{sites.pal}{optional color/fill palette for sites}

\item{sites.shape.type}{type of shape to get for points : 'solid', 'stoke', 'fill' or 'lines' (see function shape_palette for which pch is in which)}

\item{sites.size}{optional size for points (not an aes for now)}

\item{draw.arrows}{logical,draw arrows for species?}

\item{arrow.color}{color of those arrows}

\item{species.labs}{labels for species}

\item{species.geom}{either 'point' or 'text'}

\item{species.labs.size}{label size for species}

\item{species.labs.adjust}{adjust ddistance between label and arrow
ellipse args}

\item{ellipse.groups}{any of 'fill', 'color' or 'shape'; group(s) to draw ellipses for}

\item{ellipse}{logical, draw ellipse?}

\item{ellipse.labs}{logical, add labels for the ellipse? labels will be factors of groups}

\item{ellipse.level}{the confidence level at which to draw an ellipse (default is 0.95), or, if type="euclid", the radius of the circle to be drawn.}

\item{ellipse.show.legend}{logical, Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{ellipse.type}{what type of stat_ellipse to draw : 't', 'norm' or 'euclid'}

\item{ellipse.size}{width of the path of the ellipse drawn (size aes of stat_ellipse)}

\item{ellipse.alpha}{alpha transparency value for the fill aes of the ellipse}

\item{lty.pal}{optional palette for line type of ellipse
plot args}

\item{axis.lang}{language for axes, either 'FR' or 'EN' (for now if not 'FR' assumes it's 'EN')}

\item{draw.axis}{logical, draw the 0 axis on the plot?}

\item{sites.alpha}{alpha transparency value for the points (0 = transparent, 1 = opaque)
species/loadings args}
}
\value{
A ggplot2 object
}
\description{
Make an ordination plot using ggplot2
}
