% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggordi_base.R
\name{ggordi_base}
\alias{ggordi_base}
\title{Draw base plot with no sites or species}
\usage{
ggordi_base(x, draw.axis, species.geom, axis.labels)
}
\arguments{
\item{x}{scaled scores data.frame with columns for choices (PC1, PC2, etc.) and optional columns for shape, fill, color and label aes}

\item{draw.axis}{logical, draw the 0 axis on the plot?}

\item{species.geom}{either 'point' or 'text' (because text will take more space, must expand a bit)}

\item{axis.labels}{named list of x and y axis label}
}
\value{
A ggplot2 object
}
\description{
Draw base plot with no sites or species
}
