% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggordi_ellipse.fill.R
\name{ggordi_ellipse.fill}
\alias{ggordi_ellipse.fill}
\title{Draw a stat_ellipse polygon on plot (looks better when done after base plot; it doesn't hide the points. Makes it more compatible qith ggploty() too)}
\usage{
ggordi_ellipse.fill(
  g,
  ellipse.group,
  ellipse.show.legend,
  ellipse.type,
  ellipse.level,
  ellipse.size,
  ellipse.alpha,
  sites.pal
)
}
\arguments{
\item{g}{ggplot2 object returned from ggordi_base() / ggordi_sites() / ggordi_species}

\item{ellipse.group}{one of the ellipse.groups; either 'fill', 'color' or 'shape'; group to draw ellipses for}

\item{ellipse.show.legend}{logical, Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{ellipse.type}{what type of stat_ellipse to draw : 't', 'norm' or 'euclid'}

\item{ellipse.level}{the confidence level at which to draw an ellipse (default is 0.95), or, if type="euclid", the radius of the circle to be drawn.}

\item{ellipse.size}{width of the path of the ellipse drawn (size aes of stat_ellipse)}

\item{ellipse.alpha}{alpha transparency value for the fill aes of the ellipse}

\item{sites.pal}{optional color/fill palette for sites}
}
\value{
A ggplot2 object
}
\description{
Draw a stat_ellipse polygon on plot (looks better when done after base plot; it doesn't hide the points. Makes it more compatible qith ggploty() too)
}
