% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggordi_ellipse.labs.R
\name{ggordi_ellipse.labs}
\alias{ggordi_ellipse.labs}
\title{Adds labels to ellipses at their centroids}
\usage{
ggordi_ellipse.labs(g, centroids.df, ellipse.group, sites.pal)
}
\arguments{
\item{g}{a ggplot2 object to which ellipse laels will be added}

\item{centroids.df}{a data.frame with the centroids coordinates as 'x' and 'y' and the grouping variable as 'label'}

\item{sites.pal}{optional color/fill palette for sites}
}
\value{
A ggplot2 object
}
\description{
Adds labels to ellipses at their centroids
}
