% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggordi_sites.R
\name{ggordi_sites}
\alias{ggordi_sites}
\title{Draw the scores/sites on the base plot}
\usage{
ggordi_sites(
  g,
  sites.geom,
  sites.alpha,
  sites.shape.type,
  sites.pal,
  sites.leg.title.s,
  sites.leg.title.f,
  sites.leg.title.c
)
}
\arguments{
\item{g}{ggplot2 object returned from ggordi_base()}

\item{sites.geom}{either 'point' or 'text'}

\item{sites.alpha}{alpha transparency value for the points (0 = transparent, 1 = opaque)}

\item{sites.shape.type}{type of shape to get for points : 'solid', 'stoke', 'fill' or 'lines' (see function shape_palette for which pch is in which)}

\item{sites.pal}{optional color/fill palette for sites}

\item{sites.leg.title.s}{legend title for shape}

\item{sites.leg.title.f}{legend title for fill}

\item{sites.leg.title.c}{legend title for color}

\item{sites.shape}{shape aes for sites}

\item{sites.fill}{fill aes for sites}

\item{sites.color}{color aes for sites}

\item{sites.labs}{optionnal vector oflabels for sites if different than rownames}

\item{sites.size}{optional size for points (not an aes for now)}
}
\value{
A ggplot2 object
}
\description{
Draw the scores/sites on the base plot
}
