% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggordi_species.R
\name{ggordi_species}
\alias{ggordi_species}
\title{Draw the loadings/speacies on the base plot}
\usage{
ggordi_species(
  g,
  y,
  draw.arrows,
  arrow.color,
  species.labs.size,
  species.labs.adjust,
  species.geom,
  ratio
)
}
\arguments{
\item{g}{ggplot2 object returned from ggordi_base() / ggordi_sites()}

\item{draw.arrows}{logical,draw arrows for species?}

\item{arrow.color}{color of those arrows}

\item{species.labs.size}{label size for species}

\item{species.labs.adjust}{adjust ddistance between label and arrow}

\item{species.geom}{either 'point' or 'text'}

\item{ratio}{ratio to which arrows should be scaled}
}
\value{
A ggplot2 object
}
\description{
Draw the loadings/speacies on the base plot
}
